requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
cssselect2==0.7.0
Django==5.0.6
django-cors-headers==4.4.0
django-pwa==1.1.0
django-webpush==0.3.6
django-wkhtmltopdf==3.4.0
djangorestframework==3.15.2
et-xmlfile==1.1.0
frozenlist==1.4.1
html5lib==1.1
http_ece==1.2.0
idna==3.7
lxml==5.2.2
multidict==6.0.5
numpy==2.0.0
openpyxl==3.1.4
oscrypto==1.3.0
pandas==2.2.2
pdfkit==1.0.0
pillow==10.3.0
py-vapid==1.9.1
pycparser==2.22
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
pypdf==4.2.0
PyPDF2==3.0.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywebpush==2.0.0
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.2.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
svglib==1.5.1
tablib==3.6.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
webencodings==0.5.1
whitenoise==6.7.0
xhtml2pdf==0.2.16
xmltodict==0.13.0
yarl==1.9.4
